requirements.txt
fastapi
uvicorn
langchain
langchain-groq
langchain-core
python-dotenv
pydantic
